requirements.txt
PyJWT==1.7.1
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.11
psycopg2==2.8.4
flask-bcrypt==0.7.1
